requirements.txt
aiofiles==0.4.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
autocrop==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.47.1
gunicorn
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==3.4.9.31
Pillow==7.0.0
promise==2.3
pydantic==1.4
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.23.0
Rx==1.6.1
six==1.14.0
starlette==0.12.9
-e git+https://github.com/leblancfg/test_fastapi@06fa46ac6f2bcdd61c3cb7b1b9826b8170aa9b86#egg=test_fastapi
ujson==2.0.3
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
